+++ packages/nesd/lib/ui/emulator/emulator_widget.dart
@@ -19,45 +19,48 @@ class EmulatorWidget extends ConsumerWidget {
 
     final theme = Theme.of(context);
 
-    return Stack(
-      children: [
-        Focus(
-          autofocus: true,
-          onKeyEvent: (focusNode, event) =>
-              keyboardInputHandler.handleKeyEvent(event)
-              ? KeyEventResult.handled
-              : KeyEventResult.ignored,
-          child: const FrameBufferStreamBuilder(),
-        ),
-        SafeArea(
-          child: Align(
-            alignment: Alignment.topLeft,
-            child: Padding(
-              padding: const EdgeInsets.all(8),
-              child: Theme(
-                data: theme.copyWith(
-                  iconButtonTheme: IconButtonThemeData(
-                    style: theme.iconButtonTheme.style!.copyWith(
-                      backgroundColor: WidgetStateProperty.all(
-                        Colors.black.withAlpha(150),
+    return Focus(
+      canRequestFocus: false,
+      skipTraversal: true,
+      onKeyEvent: (focusNode, event) {
+        keyboardInputHandler.handleKeyEvent(event);
+
+        // make sure no unintended actions get triggered
+        return KeyEventResult.handled;
+      },
+      child: Stack(
+        children: [
+          const Focus(autofocus: true, child: FrameBufferStreamBuilder()),
+          SafeArea(
+            child: Align(
+              alignment: Alignment.topLeft,
+              child: Padding(
+                padding: const EdgeInsets.all(8),
+                child: Theme(
+                  data: theme.copyWith(
+                    iconButtonTheme: IconButtonThemeData(
+                      style: theme.iconButtonTheme.style!.copyWith(
+                        backgroundColor: WidgetStateProperty.all(
+                          Colors.black.withAlpha(150),
+                        ),
                       ),
                     ),
                   ),
-                ),
-                child: IconButton(
-                  key: menuKey,
-                  icon: const Icon(Icons.menu),
-                  color: Colors.white,
-                  onPressed: () =>
-                      ref.read(routerProvider).navigate(const MenuRoute()),
+                  child: IconButton(
+                    key: menuKey,
+                    icon: const Icon(Icons.menu),
+                    color: Colors.white,
+                    onPressed: () =>
+                        ref.read(routerProvider).navigate(const MenuRoute()),
+                  ),
                 ),
               ),
             ),
           ),
-        ),
-        if (settings.showDebugOverlay) const DebugOverlay(),
-        if (settings.showTouchControls) const TouchControlsBuilder(),
-      ],
+          if (settings.showDebugOverlay) const DebugOverlay(),
+          if (settings.showTouchControls) const TouchControlsBuilder(),
+        ],
+      ),
     );
   }
 }