+++ packages/nesd/test/helpers/host_arch_web.dart
@@ -1,0 +1,2 @@
+String hostArch() =>
+    throw UnsupportedError('Native test libraries are not used on the web');